Code review: update getHashesForGlobsAsync function
