+++ libraries/rush-lib/src/logic/operations/ShellOperationRunner.ts
@@ -29,7 +29,7 @@ import { OperationError } from './OperationError';
 import { IOperationRunner, IOperationRunnerContext } from './IOperationRunner';
 import { ProjectLogWritable } from './ProjectLogWritable';
 import { ProjectBuildCache } from '../buildCache/ProjectBuildCache';
-import { AdditionalFilesHasher } from '../buildCache/AdditionalFilesHasher';
+import { getHashesForGlobsAsync } from '../buildCache/getHashesForGlobsAsync';
 import { IOperationSettings, RushProjectConfiguration } from '../../api/RushProjectConfiguration';
 import { CollatedTerminalProvider } from '../../utilities/CollatedTerminalProvider';
 import { RushConstants } from '../RushConstants';
@@ -452,11 +452,10 @@ export class ShellOperationRunner implements IOperationRunner {
                 }
               }
               if (operationSettings.dependsOnAdditionalFiles) {
-                const additionalFiles: Map<string, string> =
-                  await AdditionalFilesHasher.getFileHashedFilesMap(
-                    operationSettings.dependsOnAdditionalFiles,
-                    this._rushProject.projectFolder
-                  );
+                const additionalFiles: Map<string, string> = await getHashesForGlobsAsync(
+                  operationSettings.dependsOnAdditionalFiles,
+                  this._rushProject.projectFolder
+                );
 
                 terminal.writeDebugLine(
                   `Including additional files to calculate build cache hash:\n  ${Array.from(
